pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "1p-si2-backend"
version = "0.1.0"
description = "Sistema de Gestión de Talleres Mecánicos - Backend API"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "1P-SI2 Team", email = "team@1p-si2.com"},
]
keywords = ["fastapi", "workshop", "management", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
]

dependencies = [
    # Core framework
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    
    # Database
    "sqlalchemy>=2.0.36,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "alembic>=1.13.0,<2.0.0",
    
    # Validation and settings
    "pydantic-settings>=2.6.1,<3.0.0",
    "pydantic[email]>=2.9.0,<3.0.0",
    
    # Authentication
    "PyJWT>=2.9.0,<3.0.0",
    
    # Email support
    "aiosmtplib>=3.0.0,<4.0.0",
    "httpx>=0.27.0,<1.0.0",
    
    # Security
    "slowapi>=0.1.9,<1.0.0",
    "starlette>=0.37.2,<0.39.0",
    "pyotp>=2.9.0,<3.0.0",
    
    # Logging
    "structlog>=24.0.0,<25.0.0",
    
    # System monitoring
    "psutil>=5.9.0,<6.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    
    # Code quality
    "ruff>=0.1.0,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    
    # Security
    "pip-audit>=2.6.0,<3.0.0",
    "bandit>=1.7.0,<2.0.0",
]

prod = [
    # Production monitoring
    "prometheus-client>=0.19.0,<1.0.0",
    "opentelemetry-api>=1.20.0,<2.0.0",
    "opentelemetry-sdk>=1.20.0,<2.0.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0,<1.0.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0,<1.0.0",
]

[project.urls]
Homepage = "https://github.com/1p-si2/backend"
Repository = "https://github.com/1p-si2/backend.git"
Issues = "https://github.com/1p-si2/backend/issues"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused-function-argument
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q", # flake8-quotes
    "PL", # pylint
    "PT", # flake8-pytest-style
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "PLR0913", # too many arguments to function call
    "PLR0915", # too many statements
    "PLR2004", # magic value used in comparison
]

[tool.ruff.per-file-ignores]
"tests/**/*" = ["PLR2004", "S101", "TID252"]
"alembic/**/*" = ["I001"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "slowapi.*",
    "pyotp.*",
    "aiosmtplib.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "venv"]
skips = ["B101", "B601"]